Consolidate exceptions and visibility
