+++ container/lib/conductor/docker/engine.py
@@ -4,7 +4,7 @@ from __future__ import absolute_import
 import logging
 plainLogger = logging.getLogger(__name__)
 
-from ..visibility import getLogger
+from container.common.visibility import getLogger
 logger = getLogger(__name__)
 
 import base64
