--- container/lib/conductor/exceptions.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import absolute_import
-
-import logging
-
-logger = logging.getLogger(__name__)
-
-
-class AnsibleContainerConductorException(Exception):
-    pass
