+++ container/lib/conductor/utils.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import
 
-from .visibility import getLogger
+from container.common.visibility import getLogger
 logger = getLogger(__name__)
 
 import os
@@ -22,7 +22,7 @@ try:
 except ImportError:
     HAS_ANSIBLE = False
 
-from .exceptions import AnsibleContainerConductorException
+from container.common.exceptions import AnsibleContainerConductorException
 
 class MakeTempDir(object):
     temp_dir = None
